Move bubble sort to a specific namespace
